Refactor: 메인 레이아웃 구조 리팩토링
- 공통 페이지 레이아웃 적용
