+++ src/components/features/Layout/MainLayout/MemberSection/index.tsx
@@ -1,5 +1,3 @@
-import { Box } from '@chakra-ui/react'
-
 import { useMemberTypeStore } from '@/stores/member-type'
 
 import { FriendSection } from './FriendSection'
@@ -9,9 +7,9 @@ export const MemberSection = () => {
   const memberType = useMemberTypeStore((state) => state.memberType)
 
   return (
-    <Box background="white" width="200px" height="full" color="text">
+    <>
       {memberType === 'FRIEND' && <FriendSection />}
       {memberType === 'KAKAO' && <SelectFriendSection />}
-    </Box>
+    </>
   )
 }
