+++ src/components/features/Layout/MainLayout/index.tsx
@@ -1,0 +1,13 @@
+import { PageLayout } from '@/components/common/PageLayout/index'
+
+import { GroupSection } from './GroupSection'
+import { MemberSection } from './MemberSection'
+
+export const MainLayout = () => {
+  return (
+    <PageLayout
+      LeftSection={<GroupSection />}
+      RightSection={<MemberSection />}
+    />
+  )
+}
