--- src/components/features/Layout/SideNavigation/index.tsx
@@ -1,5 +1,0 @@
-import { Box } from '@chakra-ui/react'
-
-export const SideNavigation = () => {
-  return <Box background="brown.300" width="40px" height="full" />
-}
